pyproject.toml
[project]
name = "vector-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "numpy>=2.2.6",
    "ruff>=0.11.11",
]
